requirements.txt
alembic==1.4.3
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chromedriver-binary==74.0.3729.6.0
click==7.1.2
colorama==0.4.3
cymem==2.0.4
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
docutils==0.16
email-validator==1.1.1
embedding-as-service-client==1.0.0
entrypoints==0.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==2.0.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython @ file:///opt/concourse/worker/volumes/live/27d08825-4fae-44a4-4324-8338f92c3ef3/volume/ipython_1604101240866/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi @ file:///opt/concourse/worker/volumes/live/152cd167-7b79-4fbd-5c97-d7b338805c2b/volume/jedi_1598371617305/work
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==21.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
pkginfo==1.5.0.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.4
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
pymongo==3.11.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.1.2
pytest-json==0.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.4
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
readme-renderer==26.0
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
-e git+https://github.com/jingwoo4710/ds-sc31x.git@f7a4997cd06736e934b9f99bc0f47def97adbe03#egg=sc31x_local
-e git+https://github.com/jingwoo4710/ds-sc32x.git@dda9a400e814193e458bab97775605ef27de204d#egg=sc32x_local
scikit-learn==0.23.2
scipy==1.5.4
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
spacy==2.3.2
SQLAlchemy==1.3.20
srsly==1.0.4
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.50.1
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
tweepy==3.9.0
tweety==0.1.6
twine==3.2.0
urllib3==1.24.3
vaderSentiment==3.3.2
wasabi==0.8.0
watchdog==0.10.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zipp==3.4.0
